mtd: Add mtd test files
This commit adds mtd test files to the bist and verifies the read
and write functionality of the mtd User partition using mtd-utils
package. The test writes random data to the mtd User partition
and reads it back to ensure that the data is same, and asserts the
test based on the comparison result. This improves the reliability
and quality of the mtd functionality.