i2c: Add i2c test files
This commit includes i2c test files to the BIST. The purpose of
the test is to check for the presence of i2c devices on the
specified bus from configuration file. To accomplish this, the
test executes "i2cdetect" command to obtain i2c bus information
from the device and then proceeds to detect all the devices listed
for the bus using periphery i2c module.