mtd: Add QSPI read and write performance tests
This commit includes MTD read and write performance tests to the
BIST. The purpose of the test is to verify the performance of QSPI
by performing read and write transaction to the MTD User partition
using 'dd' command-line utility. The test offers Read, Write and
Read_Write (optional suite) modes for performance verification.