spi: Add SPI test files for Torque sensor
This commit adds SPI Torque ADC test files to the BIST. The
purpose of the test is to detect the Torque sensor AD7797 on
the SPI bus on loading the BIST firmware. The detection is
done with the help of two separate tests by dynamic device path
lookup. The run_ad7797_id_test() verifies the Torque sensor
AD7797 ID and run_ad7797_temperature_test() reads the raw value of
inbuilt temperature sensor to verify the torque sensor presence.
The tests use python-periphery spi module to perform SPI
transactions.