can: Add CAN BIST files
This commit adds CAN test files to BIST. The purpose of the test is
to verify the functionality of CAN port (CAN_L, CAN_H and GND pins)
on the KD240 starter kit. The verification is done by performing
transmit and receive transactions between MCP25625 PMOD CAN controller
(connected via SPI to KD240) and CAN port on the kit. The test uses
python-can module for sending and receiving CAN messages.