test: tty: Add recursive search for TTY device
Ubuntu 24.04 Image introduced a change where the tty directory sits
two level deeper. It still remains in the same path in 22.04 Image.
To ensure backward compatibility, updated the code to perform a
recursive search for the TTY device. This ensures robust detection of the
TTY device as long as serial_path remains the same.

Signed-off-by: Sagar Karmarkar <sagar.karmarkar@amd.com>