tests: Add a common timeout_handler.py to handle user input timeouts
Repalacing pip package 'inputimeout' with 'signal' since inputimeout is
deprecated or unavailable in Ubuntu Noble 24.04. Instead using the signal
library module built-in in Python that provides mechanism for handling events
like user input timeouts,etc. similarly like inputimeout. This timeout_handler
can be imported in tests that need a user input on a timeout to be enabled.

Signed-off-by: Sagar Karmarkar <sagar.karmarkar@amd.com>