tests: video: Increase timeout for filesink tests to avoid test timing out
Additional warning artifact is seen due to different version of gstreamer
on Ubuntu 24.04 Noble version. This additional artifact causes filesink
tests to run into a timeout and result in failue if executed as the first
test since this artifact is a time consuming operation. Increasing the
timeout to '10' works perfectly and also provides a cushion of additional
seconds

Signed-off-by: Sagar Karmarkar <sagar.karmarkar@amd.com>