tests: gpio: Fix corner case w.r.t 1wire_brake_ctrl
Ideally, on wire connected, write_pattern and read_pattern should read
1 and 0 ie pattern inversion which is expected and should result in a
tes pass, which is the already the case now. With wire not connected
write_pattern and read_pattern are both 1 which also results in a test
pass according to the test structure now, which should not be the case.

Handled this corner case by adding a check if wp=rp and if test label is
1wire_brake_ctrl mark the test as failed with wire not connected message.

Signed-off-by: Sagar Karmarkar <sagar.karmarkar@amd.com>