docs: Added a known issue with kd240-bist FW stress testing
Loading/unloading of kd240-bist FW causes a hang on the second load of the
application. This is due to the motor modules. Workaround is to remove ADC_HUB
and all HLS modules before unloading and then user can successfully load
the FW the second time. This module removal is everytime before user
unloads kd240-bist FW.

Signed-off-by: Sagar Karmarkar <sagar.karmarkar@amd.com>