docs: Add known issue related to XXV IP
A synchronous abort kernel trace is observed sometimes which is a known
issue because of the XXV IP error. The workaround is to remove the phylink
in the kernel source code. Current workaround for BIST is to reboot/power cycle
target KR260 if this is seen while testing kr260-bist firmware

Signed-off-by: Sagar Karmarkar <sagar.karmarkar@amd.com>